Move EventMessage to actor
