+++ src/application/hooks/useActorEvent.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { useEffect, useLayoutEffect, useRef } from "react";
-import type { EventMessage } from "../../extension/messages";
 import { getPanelActor } from "../../extension/devtools/panelActor";
+import type { EventMessage } from "../../extension/actor";
 
 export function useActorEvent<TName extends EventMessage["type"]>(
   name: TName,
